requirements.txt
python-dotenv>=1.0.0
tweepy==4.14.0
openai>=1.40.8
py-cord>=2.6.1
pyyaml>=6.0.1
toml>=0.10.2
flask[async]>=2.0.0
psycopg2-binary>=2.9.9
pgvector>=0.2.3
scikit-learn>=1.3.2
numpy>=1.26.3
tenacity==8.5.0
aiohttp==3.9.1
boto3==1.35.1
Web3==7.6.1
aiohttp==3.9.1
python-telegram-bot>=20.0
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
loguru==0.7.3
tiktoken>=0.5.0
firecrawl-py>=1.11.1
typer[all]>=0.9.0
rich>=13.7.0
prompt-toolkit>=3.0.0
duckduckgo-search==7.4.3
smolagents==1.9.2
